Just set branch=tag when building a tag on Travis
Travis does this anyway, so this is removing the special-casing.
Now when we use the branch name to publish,
we actually publish into the correct folder.

0.3.1 would publish to HEAD,
as the Git fallback check got that as the branch name.
This is more correct, but this behavior is easier.

Once we want to support more CI options,
a more permanent solution will be required.
CI-Detective is a draft crate I have made,
which will help with this progress.